+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Version History
 
-## Unreleased
+## 8.0.2
 
+* [THORN-289] Add ResetCspIngest command to SpsStation. Improve message in SetCspIngest. Add cspIngestConfig attribute to SpsStation.
 * [THORN-298] Fix issue where last attribute was being marked as invalid
 * [SKB-872] Fix the response code for a number of commands that we incorrectly reporting ResultCode.OK when failing
 * [SKB-999] fix the mapping of fitted preADUs to ADC channels when verifying preADU attenuation hardware readback.